refactor: extract prettyDomain helper into util file
