+++ src/components/Popups/SelectCredentialsPopup.jsx
@@ -13,23 +13,7 @@ import { truncateByWords } from '@/functions/truncateWords';
 import { MdFactCheck } from "react-icons/md";
 import { useCredentialName } from '@/hooks/useCredentialName';
 import i18n from '@/i18n';
-
-const prettyDomain = (raw) => {
-	if (!raw) return '';
-	let value = raw.trim();
-
-	// Strip known prefixes
-	if (value.startsWith('origin:')) value = value.slice(7);
-	if (value.startsWith('x509_san_dns:')) value = value.slice(13);
-
-	// Try to reduce to hostname if it's a URL
-	try {
-		const url = new URL(value);
-		return url.host || value;
-	} catch {
-		return value;
-	}
-};
+import prettyDomain from '@/utils/prettyDomain';
 
 const SelectableCredentialSlideCard = ({
 	vcEntity,
